Introduce test manager
This commit provides a custom manager for testing clockwork. This
manager inherits all the functionality from the actual clockwork
manager, but provides custom behavior for the looping mechanism it uses
in the run method.

This change makes it so that a number of times that the clock should
tick can be specified, or the clock can be made to stop once a certain
point in time has been reached.

This allows for a clockwork manager to run in tests and not have it
operate infinitely.