Override run method
This modification allows for the max_ticks or end_time to be specified
in an options hash passed in to the run method of an instance of the
clockwork manager. These values will supercede any that were set in the
initialization of the object, before passing on to the clockwork
manager's implementation of run().