Stop requiring all files in spec_helper
This commit removes requires from the spec_helper for all files.
Instead, the file under test must be required by the spec itself. This
is to bring to light issues where classes aren't being required in the
implementation that are needed. If all the classes are loaded in the
spec_helper, this doesn't come up.

The example of this is the JobHistory class in the Manager class.

This may be mitigated in the future by moving all requires into the
Clockwork::Test module and requiring that, but at this time, it seemed
wrong to have the spec load up each file so that specs would pass when
the implementation would not actually work in a REPL without the
additional require.