Introduce test event
This commit creates an event class that inherits behavior from the
Clockwork::Event class, with the difference being that it does not
actually run.

To compensate for this, the manager now has awareness of what the job
would have run, so that may be tested if necessary without the
implementation being run otherwise.