Introduce class methods for public API surface
This commit introduces the mechanism through which external callers will
interact with this gem. Much like clockwork itself, class methods on the
module exist that will run the manager and clear it. There are also
methods exposed to be able to ascertain the job history. These
historical methods are what should be used for testing assertions
against the clockwork file under test.