Specify start time for the clockwork run
This modification allows for the tests to be run at a particular start
time that is different from the current time that the test is run. If
the start_time option is utilized, the value of Time.now is modified for
the duration of the test and returned to the proper value once the run
method has concluded.