Speed up execution runs
This modification adds in another parameter that may be passed in to the
execution run which will fast-forward through time on each clock tick at
the interval specified. This value should be set, at a maximum, to the
lowest granularity needed for any job to run. If a job is set to run
every minute, or on a particular minute, that should be the fastest
speed that is chosen. Any higher value will result in the number of
times the job is run, or if it is run at all, to be unreliable and
potentially incorrect.