Pass clockwork config to Clockwork::Test
This modification ensures that the loading of the clock file will pass
any configuration options set in to Clockwork::Test::Manager.

Previously, when the clock.rb was loaded, the config block was ignored.
This presents an issue when details in the configuration are required to
know if a job should run or not - most prominently with the time zone.

This manifested itself with events not having a timezone configured
when a default TZ was defined in the config. Tests against jobs that
should run at certain times would fail, because the time passed in would
not be parsed at the appropriate timezone in the event, and the event
would assert that it should not run when it should, were the proper
global timezone from the config passed to it.

Test coverage has been added with an acceptance test checking if a job
set to run at a certain time, in a specific timezone associated with the
global config, correctly runs.