Avoid changing global time zone state.
By assigning `Time.zone` in a `before`, we're changing the global state
of which time zone is in use for any other code which relies on it. This
should avoid that by instead relying on the block form of `use_zone`.

Note that this only works because of the slightly quirky way that test
examples save state in the `Clockwork::Test` subject, and the way that
`Clockwork::Test.run` is the thing running the relevant example. In an
ordinary test setup, the `use_zone` block be called, _then_ the RSpec
example executed. That would require the use of

  around(:each) do |example|
    orig_zone = Time.zone
    Time.zone = time_zone

    example.run

    Time.zone = orig_zone
  end

It isn't needed here though, so `use_zone` should be fine, and it has
less bookkeeping.