Handle change in run loop
This addresses a change in how the `run` method in Clockwork::Manager
operates, which was introduced in Clockwork version 2.0.1. Instead of
looping, it will now run in a thread until it is told to stop by
trapping signals.

This requires a change to the way that Clockwork::Test interacts with
the internals of Clockwork, as it was dependent on that prior internal
implementation.

This change bifurcates the processing, so that the gem will continue to
work for earlier versions of Clockwork, while adding support for 2.0.1+.