Replace Travis with GitHub Actions (#40)
This removes usage of Travis for running tests in a CI environment to
instead use GitHub Actions.

This adds an updated matrix of Ruby versions that are tested to be those
that are still supported by Ruby. That is, the lowest version of Ruby
that will be tested against as of now is 3.0.